requirements_esrgan.txt
opencv-python>=4.5.0
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
numpy>=1.19.0
realesrgan>=0.3.0
basicsr>=1.4.0 